REF: Move mongoquery and sqlite MDS into headersource. Tests pass unchanged.
